Change remote theme ref
From now on, the master branch of the profile repository will be live,
i.e. effectively deployed in the same way that the master branch of
the sandervoerman.github.io repository is the deployed version.

By contrast, the minima-profile repository will be used for
development and upstream tracking. Its master branch will track the
upstream master branch and should not be used to setup PRs or land
downstream commits for deployment.