Remove unnecessary Gemfile.lock from master branch
The master branch is deployed on Github Pages which does not make
use of the Gemfile and Gemfile.lock. Because Gemfile.lock becomes
outdated, Github starts to generate security warnings for the
outdated versions specified in the file even while those are not
the versions actually running on the Github Pages server.